Initial GitHub code repos support
Highlights:

- On app creation, a new GitHub repo is requested to be created.
- A single `Resource` model (and thus database table) for storing info, status and metadata about resources (uses single-table inheritance to provide specialised behaviour).
- Background workers for handling create and delete requests.
- Internal GitHub "agent" for talking to the GitHub API (via GitHub Apps authentication).
- Updated the `git_hub` provider spec to take into account the new config required for GitHub Apps.
- UI for showing GitHub repos and their status and info.
- Activity stream updates to show create requests, lifecycle updates, delete requests and deletion (via audits).
- Refactored the slug behaviour into a `slugged_attribute` mixin, so we can use it for both apps and resources.

Note: this PR makes little effort to properly modularise the provider specific resource logic – this is something we can do later once we better understand how different providers and resource types will actually work, and when we find the commonalities. Also, there is a lot of provider specific "stuff" across the board, from UI, backend, workers, etc., so this will need some proper thinking on the best way to modularise and make it easy for anyone to add a new provider.