Remove auto prefixing of resource names
Previously, any new resource created would have it's name prefixed with the slug of the project (aka space) that resource was contained. This was intended to isolate resources to avoid name clashes between projects/spaces.

However, given the current open projects/spaces model, this prefixing was unnecessary complexity and inflexibility at this stage. We can always add this back in for closed teams/projects/space if/when we get to that.