Gracefully handle deletion of GitHub code repos that don't have a `full_name`
If the request create fails, then the resource will be in a `failed` state and without a `full_name` value (usually). This change now allows the resource entry within the hub to be deleted, without having to talk to the agent (as we don't know which repo it's actually referring to at this point).