GitHub Agent: Don't apply best practices if 'master' branch doesn't exist
Partial fix for #153.

If a repo exists, and it doesn't have a 'master` branch yet, we can't enforce the best practices. This change ensures that the create request still goes through without an error. But more work is needed to figure out the best way of enforcing the best practices later on (when the branch does exist).