Auth logout
Closes #67

A new config option is now available – `AUTH_LOGOUT_URL` – to specify the logout URL (that needs to be handled by the auth proxy).

Note: this currently doesn't work properly with the local Keycloak Gatekeeper setup – it errors with a `400` error. More investigation is required (which we can do separately).